+++ src/constants/numbers.ts
@@ -2,7 +2,7 @@
  * The default maximum number of milliseconds to wait for a promise to resolve.
  * One hour.
  */
-export const DEFAULT_MAX_TIME_PROMISE = 3_600_000;
+export const DEFAULT_MAX_TIME_PROMISE = 600_000;
 
 /**
  * The default minimum number of milliseconds to wait for an activity to complete.