+++ src/interpreters/__tests__/data/machine2.cov.test.ts
@@ -301,7 +301,7 @@ describe('machine coverage', () => {
       });
 
       describe('#35 => Close the service', async () => {
-        test('#01 => Pause the service', service.pause.bind(service));
+        test('#01 => Pause the service', service.stop);
 
         describe('#02 => Calls of log', () => {
           test('#01 => Length of calls of log is the same of length of strings', () => {