+++ src/states/functions/nodeToValue.test.ts
@@ -1,8 +1,0 @@
-import { nodeToValue } from './nodeToValue';
-
-describe('nodeToValue - coverage', () => {
-  test('#01 => Atomic', () => {
-    const out = nodeToValue({});
-    expect(out).toStrictEqual({});
-  });
-});